+++ app/api/comment_routes.py
@@ -1,4 +1,4 @@
-from flask import Blueprint, request, abort
+from flask import Blueprint, request
 from flask_login import current_user
 
 #comment model file not created yet
@@ -27,10 +27,6 @@ def post_comment():
         db.session.add(new_comment)
         db.session.commit()
         return new_comment.to_dict()
-
-    if form.errors:
-        # print(form.errors
-        return form.errors, 403
 
 @comment_routes.route('/<int:id>/', methods=["PATCH"])
 def patch_comment(id):
