+++ app/forms/comment_form.py
@@ -1,13 +1,7 @@
-from xml.dom import ValidationErr
 from flask import Flask
 from flask_wtf import FlaskForm
 from wtforms import TextAreaField, IntegerField
-from wtforms.validators import DataRequired, ValidationError
-
-# def validate_comment(field):
-#     if len(field.data) == 0:
-#         raise ValidationError('Please Input a Comment')
-
+from wtforms.validators import DataRequired
 
 class CommentForm(FlaskForm):
     body = TextAreaField('Comment', validators=[DataRequired()])
