+++ react-app/src/components/photos/PhotoFeed.js
@@ -4,8 +4,7 @@ import { editPhoto, grabPhotos } from "../../store/photo";
 import PhotoModal from './PhotoModal';
 import CommentsFeed from '../comments/CommentsFeed'
 
-// import PhotoContainer from './PhotoContainer';
-import "./photos.css"
+import PhotoContainer from './PhotoContainer';
 
 const PhotoFeed = () => {
 	const dispatch = useDispatch();
@@ -24,10 +23,10 @@ const PhotoFeed = () => {
 					<h1>This is your photo feed</h1>
 					<ul>
 						{photos.map(photo => (
-							<li key={photo.id} className="photo-li">
-								{/* <PhotoContainer photo={photo} /> */}
-								<PhotoModal  photo={photo}/>
-								<CommentsFeed photo={photo}/>
+							<li key={photo.id}>
+								<PhotoContainer photo={photo} />
+								<PhotoModal photo={photo} />
+								<CommentsFeed photo={photo} />
 							</li>
 						))}
 					</ul>
