--- react-app/src/components/photos/PhotoModal/menu.css
@@ -1,66 +1,0 @@
-#menu-toggle {
-    display: none;
-}
-
-.hamburger {
-    width: 40px;
-    height: 40px;
-    background-color: white;
-    position: relative;
-    cursor: pointer;
-    border-radius: 2px;
-}
-
-.bun {
-    width: 30px;
-    height: 30px;
-    position: absolute;
-    top: 9px;
-    left: 9px;
-    transition: transform .18s
-    cubic-bezier(0.04, 0.04,
-    0.12, 0.96) .15s;
-}
-
-.bun-crust {
-    display: block;
-    width: 17px;
-    height: 1px;
-    background-color: black;
-    border-radius: 1px;
-    position: absolute;
-    left: 7px;
-    transition: transform .1596s
-    cubic-bezier(0.52, 0.16,
-    0.52, 0.84) .2s;
-}
-
-.bun-crust-top {
-    top: 14px;
-    transform: translateY(-3px);
-}
-
-.bun-crust-bottom {
-    bottom: 14px;
-    transform: translateY(3px);
-}
-
-#menu-toggle:checked +
-.hamburger .bun-top {
-    transform: rotate(45deg);
-}
-
-#menu-toggle:checked +
-.hamburger .bun-bottom {
-    transform: rotate(-45deg);
-}
-
-#menu-toggle:checked +
-.hamburger .bun-crust-top,
-#menu-toggle:checked +
-.hamburger .bun-crust-bottom {
-    transform: none;
-    transition: transform .1806s
-    cubic-bezier(0.04, 0.04,
-    0.12, 0.96);
-}
