+++ react-app/src/context/Modal.css
@@ -15,13 +15,10 @@
     right: 0;
     left: 0;
     bottom: 0;
-    background-color: rgba(5, 5, 5, 0.7);
+    background-color: rgba(0, 0, 0, 0.7);
   }
 
   #modal-content {
     position: absolute;
-    background-color:rgb(255, 255, 255);
-    display: flex;
-    justify-content: center;
-    flex-direction: row;
+    background-color:white;
   }
