+++ react-app/src/index.css
@@ -1,6 +1,1 @@
 /* TODO Add site wide styles */
-body{
-  margin-top: 10px;
-  background-color: lightslategray;
-
-}
