+++ react-app/src/index.js
@@ -5,7 +5,6 @@ import './index.css';
 import App from './App';
 import configureStore from './store';
 import { ModalProvider } from './context/Modal';
-import "./index.css"
 
 const store = configureStore();
